requirements.txt
PyGithub
github3.py
click
requests
tabulate
pyyaml
